+++ src/Components/NavTree/TypesNavTree.jsx
@@ -1,159 +1,70 @@
-import clsx from 'clsx'
-import React, {ReactElement, forwardRef} from 'react'
-import PropTypes from 'prop-types'
-import TreeItem, {useTreeItem} from '@mui/lab/TreeItem'
-import Box from '@mui/material/Box'
-import Typography from '@mui/material/Typography'
-import HideToggleButton from '../HideToggleButton'
+import React, {ReactElement, RefObject, forwardRef, useRef} from 'react'
+import TreeItem from '@mui/lab/TreeItem'
+import CustomContent from './CustomContent'
 import NavTree from './NavTree'
+import PropTypes from './PropTypes'
 
 
 /**
  * @property {object} model IFC model
+ * @property {object} element Element in the model
  * @property {object} types Types to use in the model
  * @property {string} pathPrefix URL prefix for constructing links to
  *   elements, recursively grown as passed down the tree
  * @property {Function} selectWithShiftClickEvents handler for shift-clicks
+ * @property {Map<string,RefObject<HTMLDivElement>>} idToRef Mapping of expressId to TreeItem refs
  * @return {ReactElement}
  */
 export default function TypesNavTree({
+  keyId,
   model,
+  element,
   types,
   pathPrefix,
   selectWithShiftClickEvents,
+  idToRef,
 }) {
-  const CustomContent = forwardRef(function CustomContent(props, ref) {
-    const {
-      classes,
-      className,
-      label,
-      nodeId,
-      icon: iconProp,
-      expansionIcon,
-      displayIcon,
-      hasHideIcon,
-    } = props
-
-    const {
-      disabled,
-      expanded,
-      selected,
-      focused,
-      handleExpansion,
-      handleSelection,
-      preventSelection,
-    } = useTreeItem(nodeId)
-
-    const icon = iconProp || expansionIcon || displayIcon
-
-    const handleMouseDown = (event) => preventSelection(event)
-
-    const handleExpansionClick = (event) => handleExpansion(event)
-
-    const handleSelectionClick = (event) => {
-      handleSelection(event)
-    }
-
-    return (
-      // eslint-disable-next-line jsx-a11y/no-static-element-interactions
-      <div
-        className={clsx(className, classes.root, {
-          [classes.expanded]: expanded,
-          [classes.selected]: selected,
-          [classes.focused]: focused,
-          [classes.disabled]: disabled,
-        })}
-        onMouseDown={handleMouseDown}
-        ref={ref}
-      >
-        <Box
-          onClick={handleExpansionClick}
-          sx={{margin: '0px 14px 0px 14px'}}
-        >
-          {icon}
-        </Box>
-        <div style={{width: '300px'}}>
-          <Typography
-            variant='tree'
-            onClick={handleSelectionClick}
-          >
-            {label}
-          </Typography>
-          {hasHideIcon &&
-            <div style={{display: 'contents'}}>
-              <HideToggleButton elementId={nodeId}/>
-            </div>
-          }
-        </div>
-      </div>
-    )
-  })
-
-  CustomContent.propTypes = TypesNavTreePropTypes
+  const customContentRef = forwardRef(CustomContent)
+  customContentRef.propTypes = PropTypes
 
   const CustomTreeItem = (props) => {
-    return <TreeItem ContentComponent={CustomContent} {...props}/>
+    return <TreeItem ContentComponent={customContentRef} {...props}/>
   }
 
-  let i = 0
+  // TODO(pablo): total hack to support scrollIntoView behavior.  See
+  // NavTreePanel#useEffect[selectedElts] for use.
+  const itemRef = useRef(null)
+  const nodeId = element.expressID.toString()
+  const itemId = `type-${nodeId}`
+  idToRef[itemId] = itemRef
 
+  let i = 0
   return types.map((type) =>
     <CustomTreeItem
-      key={type.name}
-      nodeId={type.name}
+      key={keyId}
+      nodeId={itemId}
       label={type.name}
       ContentProps={{
         hasHideIcon: type.elements && type.elements.length > 0,
       }}
     >
+      <div ref={itemRef}/>
       {type.elements && type.elements.length > 0 ?
-    type.elements.map((e) => {
-      const childKey = `${pathPrefix}-${i++}`
-      return (
-        <NavTree
-          key={childKey}
-          model={model}
-          element={e}
-          pathPrefix={pathPrefix}
-          selectWithShiftClickEvents={selectWithShiftClickEvents}
-        />
-      )
-    }) : null}
+       type.elements.map((e) => {
+         const childKeyId = `${pathPrefix}-${i++}`
+         return (
+           <NavTree
+             keyId={childKeyId}
+             model={model}
+             element={e}
+             pathPrefix={pathPrefix}
+             selectWithShiftClickEvents={selectWithShiftClickEvents}
+             idToRef={idToRef}
+           />
+         )
+       }) : null
+      }
     </CustomTreeItem>)
 }
 
 
-const TypesNavTreePropTypes = {
-  /**
-   * Override or extend the styles applied to the component.
-   */
-  classes: PropTypes.object.isRequired,
-  /**
-   * className applied to the root element.
-   */
-  className: PropTypes.string,
-  /**
-   * The icon to display next to the tree node's label. Either a parent or end icon.
-   */
-  displayIcon: PropTypes.node,
-  /**
-   * The icon to display next to the tree node's label. Either an expansion or collapse icon.
-   */
-  expansionIcon: PropTypes.node,
-  /**
-   * The icon to display next to the tree node's label.
-   */
-  icon: PropTypes.node,
-  /**
-   * The tree node label.
-   */
-  label: PropTypes.node,
-  /**
-   * The id of the node.
-   */
-  nodeId: PropTypes.string.isRequired,
-  /**
-   * Determines if the tree node has a hide icon.
-   */
-  hasHideIcon: PropTypes.bool,
-}
